extract parsed description to component
